move date formatting to util and refactor name to title
