+++ bundles/framework/announcements/view/AnnouncementsCollapse.jsx
@@ -2,22 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Message, Collapse, CollapsePanel, Divider } from 'oskari-ui';
 import { AnnouncementsContent } from './';
+import { getDateRange } from '../service/util';
 
 // Collapse panel -> set title, locale and date range according to announcement
-
-const TIME_OPTIONS = {
-    hour: '2-digit',
-    minute: '2-digit'
-};
-const RANGE_SEPARATOR = '\u2013';
-
-const formatDate = (isoDateTime) => {
-    const dateTime = new Date(isoDateTime);
-    const date = dateTime.toLocaleDateString();
-    const time = dateTime.toLocaleTimeString([], TIME_OPTIONS);
-
-    return `${time} ${date}`;
-};
 
 export const AnnouncementsCollapse = ({
     announcements
@@ -31,12 +18,10 @@ export const AnnouncementsCollapse = ({
         <div>
             <Collapse accordion>
                 { announcements.map((announcement) => {
-                    const { name } = Oskari.getLocalized(announcement.locale);
-                    const start = formatDate(announcement.beginDate);
-                    const end = formatDate(announcement.endDate);
-                    const dateRange = start + RANGE_SEPARATOR + end;
+                    const { title } = Oskari.getLocalized(announcement.locale);
+                    const dateRange = getDateRange(announcement);
                     return (
-                        <CollapsePanel header={name} key={announcement.id}>
+                        <CollapsePanel header={title} key={announcement.id}>
                             <AnnouncementsContent announcement={announcement}/>
                             <Divider />
                             <b><Message messageKey={'valid'} /></b>
